Padronização do código (Squash)
Retirando algumas duplicações de codigo

Começando a reduzir a duplicacao de codigo com métodos

Transformando duplicação de código em métodos

Update README.md

Removendo mais codigos duplicados e os transformando em métodos

Tirando a duplicação restante da analises.py com a app.py

Tirando duplicação de código em db_functions

Tirando mais métodos duplicados de analises.py e app.py

Tirando mais um método duplicado de analises.py

Diminuindo mais código duplicado

Subindo .codeclimate.yml

Removendo o codeclimate

Tirando mais métodos

Tirando mais métodos

Começando a modularização do projeto editor

Aplicando a tecnica Single Responsiblity principle e modularizando o código de login

Aplicando a tecnica Single Responsiblity principle modularizando as endpoints de configuracoes

Aplicando a técnica Single Responsability principle modularizando as endpoints de upload

Aplicando a técnica Single Responsability principle aos arquivos app.py e pendencias.py

 Movendo mapa_pendencias para o modulo configuracoes

Aplicando tecnica Single Responsiblity principle modularizando o dowload

Update README.md

Modularizando mais o projeto

Modularizando mais o projeto
